feat(transactions): add balance checks and update logic for transaction creation, update, and deletion

+++ src/transactions/transactions.repository.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateTransactionDto } from './dto/create-transaction.dto';
 import { UpdateTransactionDto } from './dto/update-transaction.dto';
 import { PrismaService } from '../prisma/prisma.service';
@@ -18,9 +22,44 @@ export class TransactionsRepository {
   }
 
   async create(createTransactionDto: CreateTransactionDto) {
-    return this.prisma.transaction.create({
-      data: createTransactionDto,
+    const account = await this.prisma.account.findUnique({
+      where: { id: createTransactionDto.accountId },
     });
+
+    if (!account) {
+      throw new NotFoundException('Account not found');
+    }
+
+    // Pengubahan saldo dari desimal ke number untuk perbandingan
+    const balance: number = Number(account.balance ?? 0);
+
+    // Cek saldo untuk tipe pengeluaran
+    if (
+      (createTransactionDto.type === 'EXPENSE' ||
+        createTransactionDto.type === 'TRANSFER') &&
+      balance < createTransactionDto.amount
+    ) {
+      throw new BadRequestException('Insufficient funds');
+    }
+
+    if (createTransactionDto.type === 'INCOME') {
+      return await this.prisma.$transaction([
+        this.prisma.transaction.create({ data: createTransactionDto }),
+        this.prisma.account.update({
+          where: { id: createTransactionDto.accountId },
+          data: { balance: { increment: createTransactionDto.amount } },
+        }),
+      ]);
+    }
+
+    // Default return untuk pengeluaran
+    return await this.prisma.$transaction([
+      this.prisma.transaction.create({ data: createTransactionDto }),
+      this.prisma.account.update({
+        where: { id: createTransactionDto.accountId },
+        data: { balance: { decrement: createTransactionDto.amount } },
+      }),
+    ]);
   }
 
   async update(id: number, updateTransactionDto: UpdateTransactionDto) {
@@ -30,10 +69,31 @@ export class TransactionsRepository {
       throw new NotFoundException('Transaction not found');
     }
 
-    return this.prisma.transaction.update({
-      where: { id },
-      data: updateTransactionDto,
-    });
+    // Transaksi untuk pemasukan
+    if (transaction.type === 'INCOME') {
+      return await this.prisma.$transaction([
+        this.prisma.transaction.update({
+          where: { id },
+          data: updateTransactionDto,
+        }),
+        this.prisma.account.update({
+          where: { id: transaction.accountId },
+          data: { balance: { increment: updateTransactionDto.amount ?? 0 } },
+        }),
+      ]);
+    }
+
+    // Transaksi untuk pengeluaran
+    return await this.prisma.$transaction([
+      this.prisma.transaction.update({
+        where: { id },
+        data: updateTransactionDto,
+      }),
+      this.prisma.account.update({
+        where: { id: transaction.accountId },
+        data: { balance: { decrement: updateTransactionDto.amount ?? 0 } },
+      }),
+    ]);
   }
 
   async remove(id: number) {
@@ -43,8 +103,28 @@ export class TransactionsRepository {
       throw new NotFoundException('Transaction not found');
     }
 
-    return this.prisma.transaction.delete({
-      where: { id },
-    });
+    // Transaksi untuk pemasukan
+    if (transaction.type === 'INCOME') {
+      return await this.prisma.$transaction([
+        this.prisma.transaction.delete({
+          where: { id },
+        }),
+        this.prisma.account.update({
+          where: { id: transaction.accountId },
+          data: { balance: { decrement: transaction.amount } },
+        }),
+      ]);
+    }
+
+    // Transaksi untuk pengeluaran
+    return await this.prisma.$transaction([
+      this.prisma.transaction.delete({
+        where: { id },
+      }),
+      this.prisma.account.update({
+        where: { id: transaction.accountId },
+        data: { balance: { increment: transaction.amount } },
+      }),
+    ]);
   }
 }
